pyproject.toml
[tool.poetry]
name = "mujoco-logger"
version = "0.1.0"
description = ""
authors = ["Lev Kozlov <kozlov.l.a10@gmail.com>"]
license = "GNU General Public License v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mujoco = "^3.1.1"
numpy-quaternion = "^2023.0.2"

[tool.poetry.extras]
examples = ["robot_descriptions", "matplotlib"]

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
line-length = 120

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
